+++ src/valve_sdk/netvars.hpp
@@ -13,18 +13,18 @@ private:
 	struct stored_data
 	{
 		RecvProp* prop_ptr;
-		uint16_t class_relative_offset;
+		uint32_t class_relative_offset;
 	};
 
 public:
 	// Return a const instance, others shouldnt modify this.
-	static auto get() -> const netvar_manager &
+	static auto get() -> const netvar_manager&
 	{
 		static netvar_manager instance;
 		return instance;
 	}
 
-	auto get_offset(const fnv::hash hash) const -> std::uint16_t
+	auto get_offset(const fnv::hash hash) const -> std::uint32_t
 	{
 		return m_props.at(hash).class_relative_offset;
 	}
@@ -34,15 +34,15 @@ public:
 		return m_props.at(hash).prop_ptr;
 	}
 
-	__declspec(noinline) static auto get_offset_by_hash(const fnv::hash hash) -> std::uint16_t
+	__declspec(noinline) static auto get_offset_by_hash(const fnv::hash hash) -> std::uint32_t
 	{
 		return get().get_offset(hash);
 	}
 
 	template<fnv::hash Hash>
-	static auto get_offset_by_hash_cached() -> std::uint16_t
+	static auto get_offset_by_hash_cached() -> std::uint32_t
 	{
-		static auto offset = std::uint16_t(0);
+		static auto offset = std::uint32_t(0);
 		if (!offset)
 			offset = get_offset_by_hash(Hash);
 		return offset;
@@ -50,7 +50,7 @@ public:
 
 private:
 	netvar_manager();
-	auto dump_recursive(const char* base_class, RecvTable* table, std::uint16_t offset) -> void;
+	auto dump_recursive(const char* base_class, RecvTable* table, std::uint32_t offset) -> void;
 
 private:
 	std::map<fnv::hash, stored_data> m_props;
